server/requirements.txt
websockets
python-slugify